Refactor the app structure
