--- navigation.py
@@ -1,33 +1,0 @@
-"""Module for navigation through images"""
-
-
-class ImageList:
-    """Handles iterating through the images.
-    """
-    def __init__(self, images: list):
-        self.image_list = images
-        self.n = -1
-        self.max = len(images)
-
-    def next(self):
-        """Sets the next image as current.
-        """
-        self.n += 1
-
-        if self.n < self.max:
-            current_image = self.image_list[self.n]
-        else:
-            self.n = 0
-            current_image = self.image_list[self.n]
-        return current_image
-
-    def prev(self):
-        """Sets the previous image as current.
-        """
-        if self.n == 0:
-            self.n = self.max - 1
-            current_image = self.image_list[self.n]
-        else:
-            self.n -= 1
-            current_image = self.image_list[self.n]
-        return current_image
